mos_6569: Cycles emulation.
- Added real cycle emulation
- Support for border hacks (enabling/disabling)
- FLD mode fixed

The display area is painted 8 pixesl per cycle as it is expected,
on the other hand, the sprites and the borders are generated at
once just before the final rendering on the last horizontal cycle.

There are some synchronisation problems between some games and the
raster irq. For this reason the raster irq is not triggered when it
is intended to (cycles 0 and 1) but delayed to cycle 10.