ASpace: Removed nuisance of entire addr_t range.
Previously, the memory mappings had to include all the set of
addresses belonging to the addr_t type even if those addresses were
not used/accessed (for example, to define the mappings of a 16K
address space, 64K had to be defined because addr_t is 16bits).
This is now changed, thanks to a new argument "amask" (address mask)
that tells ASpace which set of addresses are of interest.