New I/O Expansion devs: Cartridge and CartGeneric.
Cartridge is a device associated to a physical .crt file
(CRT format) that contains information about a cartridge.
Its purpose is to handle GAME/EXROM pins and, based on ROMH/ROML
lines, it provides the devices to handle certain memory areas
(usually ROM devices containing games or applications).

CartGeneric is an implementation of Cartridge, which supports
cartridge type 0 (Generic Cartridge), that is 8K and 16K normal
cartridges and Ultimax 4K, 8K and 16K.