ui: Workaround for SFML keyboard problems
It is known that SFML does not recognise keys from ISO keyboards
and a workaround for this was added some time ago.

Besides that, under MacOS it treats the tilde ('~') key as minus
('-') and the minus key is not recognosed at all.
This commit adds a workaround for this problem.

High priority TODO: New UI using SDL2